Formatting - fixing spaces and minor problems
