+++ Ficek_QuantiApp/Ficek_QuantiApp.swift
@@ -5,14 +5,14 @@ import RocketDetail
 
 @main
 struct Ficek_QuantiApp: App {
-    var body: some Scene {
-        WindowGroup {
-            RocketListView(
-                store: Store(
-                    initialState: RocketListCore.State(),
-                    reducer: RocketListCore()._printChanges()
-                )
-            )
-        }
+  var body: some Scene {
+    WindowGroup {
+      RocketListView(
+        store: Store(
+          initialState: RocketListCore.State(),
+          reducer: RocketListCore()._printChanges()
+        )
+      )
     }
+  }
 }
