+++ Ficek_QuantiApp/Ficek_QuantiApp/Domain/Package.swift
@@ -4,61 +4,59 @@
 import PackageDescription
 
 let package = Package(
-    name: "Domain",
-    platforms: [.iOS(.v16)],
-    products: [
-        .library(
-            name: "RocketClient",
-            targets: ["RocketClient"]),
-        .library(
-            name: "ErrorHandlingConcurrency",
-            targets: ["ErrorHandlingConcurrency"]),
-    ],
-    dependencies: [
-        .package(url: "https://github.com/pointfreeco/swift-composable-architecture", from: "0.9.0"),
-        .package(path: "../Infrastructure/"),
-        .package(url: "https://github.com/Qase/swift-core", branch: "develop"),
-    ],
-    
-    targets: [
-       .target(
-            name: "RocketClient",
-            dependencies: [
-                .product(name: "ComposableArchitecture", package: "swift-composable-architecture"),
-                .product(name: "Networking", package: "swift-core"),
-                .product(name: "NetworkClientDependency", package: "Infrastructure"),
-                "ErrorHandlingConcurrency"
-            ]
-        ),
-       .target(
-            name: "ErrorHandlingConcurrency",
-            dependencies: [
-                .product(name: "ComposableArchitecture", package: "swift-composable-architecture"),
-                .product(name: "Networking", package: "swift-core"),
-                .product(name: "NetworkClientDependency", package: "Infrastructure"),
-            ]
-        ),
-        
-        .testTarget(
-            name: "RocketClientTests",
-            dependencies: [
-                .product(name: "ComposableArchitecture", package: "swift-composable-architecture"),
-                .product(name: "Networking", package: "swift-core"),
-                .product(name: "RequestBuilder", package: "swift-core"),
-                .product(name: "NetworkMonitoring", package: "swift-core"),
-                .product(name: "ModelConvertible", package: "swift-core"),
-                .product(name: "XCTestHelper", package: "Infrastructure"),
-                "RocketClient"
-            ]
-        ),
-       .testTarget(
-           name: "RocketConvertersTests",
-           dependencies: [
-               .product(name: "ComposableArchitecture", package: "swift-composable-architecture"),
-               .product(name: "ModelConvertible", package: "swift-core"),
-               .product(name: "XCTestHelper", package: "Infrastructure"),
-               "RocketClient"
-           ]
-       ),
-    ]
+  name: "Domain",
+  platforms: [.iOS(.v16)],
+  products: [
+    .library(
+      name: "RocketClient",
+      targets: ["RocketClient"]),
+    .library(
+      name: "ErrorHandlingConcurrency",
+      targets: ["ErrorHandlingConcurrency"]),
+  ],
+  dependencies: [
+    .package(url: "https://github.com/pointfreeco/swift-composable-architecture", from: "0.9.0"),
+    .package(path: "../Infrastructure/"),
+    .package(url: "https://github.com/Qase/swift-core", branch: "develop"),
+  ],
+  targets: [
+    .target(
+      name: "RocketClient",
+      dependencies: [
+        "ErrorHandlingConcurrency",
+        .product(name: "ComposableArchitecture", package: "swift-composable-architecture"),
+        .product(name: "Networking", package: "swift-core"),
+        .product(name: "NetworkClientDependency", package: "Infrastructure"),
+      ]
+    ),
+    .target(
+      name: "ErrorHandlingConcurrency",
+      dependencies: [
+        .product(name: "ComposableArchitecture", package: "swift-composable-architecture"),
+        .product(name: "Networking", package: "swift-core"),
+        .product(name: "NetworkClientDependency", package: "Infrastructure"),
+      ]
+    ),
+    .testTarget(
+      name: "RocketClientTests",
+      dependencies: [
+        "RocketClient",
+        .product(name: "ComposableArchitecture", package: "swift-composable-architecture"),
+        .product(name: "Networking", package: "swift-core"),
+        .product(name: "RequestBuilder", package: "swift-core"),
+        .product(name: "NetworkMonitoring", package: "swift-core"),
+        .product(name: "ModelConvertible", package: "swift-core"),
+        .product(name: "XCTestHelper", package: "Infrastructure"),
+      ]
+    ),
+    .testTarget(
+      name: "RocketConvertersTests",
+      dependencies: [
+        "RocketClient",
+        .product(name: "ComposableArchitecture", package: "swift-composable-architecture"),
+        .product(name: "ModelConvertible", package: "swift-core"),
+        .product(name: "XCTestHelper", package: "Infrastructure"),
+      ]
+    ),
+  ]
 )
