+++ Ficek_QuantiApp/Ficek_QuantiApp/Domain/Sources/RocketClient/DependencyValues+rocketClient.swift
@@ -2,8 +2,8 @@ import Foundation
 import Dependencies
 
 extension DependencyValues {
-    public var rocketClient: RocketClient {
-        get { self[RocketClient.self] }
-        set { self[RocketClient.self] = newValue }
-    }
+  public var rocketClient: RocketClient {
+    get { self[RocketClient.self] }
+    set { self[RocketClient.self] = newValue }
+  }
 }
