+++ Ficek_QuantiApp/Ficek_QuantiApp/Domain/Sources/RocketClient/RocketClient+DependencyKey.swift
@@ -20,10 +20,11 @@ extension RocketClient: DependencyKey {
         let request = Request(
           endpoint: Self.RocketRequest.allRockets.rawValue
         )
+        
         return request
           .execute(using: networkClient)
           .convertToDomainModel(using: converter)
-//          .eraseToAnyPublisher()
+        //          .eraseToAnyPublisher()
       },
       fetchAsync: {
         let request = Request(
@@ -57,56 +58,3 @@ extension RocketClient: DependencyKey {
   )
 }
 
-extension NetworkError: ModelConvertibleErrorCapable {
-  public static var modelConvertibleError: NetworkError {
-    .timeoutError
-  }
-}
-
-extension NetworkError: ErrorHandlingConcurrency {
-  public var causeUIDescription: String {
-    switch cause {
-    case .invalidResponse:
-      return "Invalid response"
-    case .urlError:
-      return "Error with url"
-    case .unauthorized:
-      return "Bad credentials"
-    case .clientError:
-      return "Client error"
-    case .serverError:
-      return "Server error"
-    case .noConnection:
-      return "No connection"
-    case .jsonDecodingError:
-      return "Internal error"
-    case .urlRequestBuilderError:
-      return "Internal error"
-    case .timeout:
-      return "Timeout"
-    }
-  }
-  
-  public var causeName: String {
-    switch cause {
-    case .invalidResponse:
-      return "Invalid response"
-    case .urlError:
-      return "Url error"
-    case .unauthorized:
-      return "Unauthorized"
-    case .clientError:
-      return "Client error"
-    case .serverError:
-      return "Server error"
-    case .noConnection:
-      return "No connection"
-    case .jsonDecodingError:
-      return "Json decoding error"
-    case .urlRequestBuilderError:
-      return "Url request error"
-    case .timeout:
-      return "Timeout"
-    }
-  }
-}
