+++ Ficek_QuantiApp/Ficek_QuantiApp/Domain/Sources/RocketClient/RocketClient.swift
@@ -4,14 +4,14 @@ import Combine
 import RequestBuilder
 
 public struct RocketClient {
-    
-    public enum RocketRequest: String {
-       // private var baseURL: String { return "https://api.spacexdata.com/v4" }
-        case allRockets = "https://api.spacexdata.com/v4/rockets"
-    }
-    
-    public var fetchAllRockets: ()  -> AnyPublisher<[Rocket], RocketError>
-        
-    public var fetchAsync: () async throws -> [Rocket]
+  
+  public enum RocketRequest: String {
+    // private var baseURL: String { return "https://api.spacexdata.com/v4" }
+    case allRockets = "https://api.spacexdata.com/v4/rockets"
+  }
+  
+  public var fetchAllRockets: ()  -> AnyPublisher<[Rocket], RocketError>
+  
+  public var fetchAsync: () async throws -> [Rocket]
 }
