+++ Ficek_QuantiApp/Ficek_QuantiApp/Features/Sources/RocketDetail/Core/RocketParametersCore.swift
@@ -4,7 +4,7 @@ import Dependencies
 import RocketClient
 import RocketLaunch
 
-public struct RocketParametersCore: ReducerProtocol{
+public struct RocketParametersCore: ReducerProtocol {
   
   public init() { }
   
@@ -12,7 +12,7 @@ public struct RocketParametersCore: ReducerProtocol{
     public var id: String { rocket.id }
     public var name: String { rocket.name }
     public var rocket: Rocket
-    public var height: String {"\(Int(round(rocket.height.meters ?? 0)))m"}
+    public var height: String {"\(Int(round(rocket.height.meters ?? 0)))m" }
     public var diameter: String { "\(Int(round(rocket.diameter.meters ?? 0)))m" }
     public var mass: String {
       if let mass = rocket.mass.kg {
