+++ Ficek_QuantiApp/Ficek_QuantiApp/Features/Sources/RocketDetail/RocketPhotosView.swift
@@ -3,44 +3,43 @@ import ComposableArchitecture
 import RocketLaunch
 
 struct RocketPhotosView: View {
-    let store: StoreOf<RocketPhotosCore>
-    
-    var body: some View {
-        WithViewStore(self.store) { viewStore in
-            VStack(alignment: .leading) {
-                Text("Photos")
-                    .font(.headline)
-                    .padding(.leading, 16)
-                
-                ForEach (viewStore.rocket.flickrImages, id: \.self) { image in
-                    
-                    AsyncImage(
-                        url: URL(string: image),
-                        content: { asyncImage in
-                            asyncImage
-                                .resizable()
-                                .clipShape(RoundedRectangle(cornerRadius: 8))
-                                .frame(maxWidth: .infinity, maxHeight: .infinity)
-                                .scaledToFit()
-                                .padding(15)
-                        },
-                        placeholder: {
-                            ProgressView()
-                        }
-                    )
-                }
+  let store: StoreOf<RocketPhotosCore>
+  
+  var body: some View {
+    WithViewStore(self.store) { viewStore in
+      VStack(alignment: .leading) {
+        Text("Photos")
+          .font(.headline)
+          .padding(.leading, 16)
+        
+        ForEach (viewStore.rocket.flickrImages, id: \.self) { image in
+          AsyncImage(
+            url: URL(string: image),
+            content: { asyncImage in
+              asyncImage
+                .resizable()
+                .clipShape(RoundedRectangle(cornerRadius: 8))
+                .frame(maxWidth: .infinity, maxHeight: .infinity)
+                .scaledToFit()
+                .padding(15)
+            },
+            placeholder: {
+              ProgressView()
             }
+          )
         }
+      }
     }
+  }
 }
 
 struct RocketPhotosView_Previews: PreviewProvider {
-    static var previews: some View {
-        RocketPhotosView(
-            store: Store(
-                initialState: RocketPhotosCore.State(rocket: .mock),
-                reducer: RocketPhotosCore()
-            )
-        )
-    }
+  static var previews: some View {
+    RocketPhotosView(
+      store: Store(
+        initialState: RocketPhotosCore.State(rocket: .mock),
+        reducer: RocketPhotosCore()
+      )
+    )
+  }
 }
