+++ Ficek_QuantiApp/Ficek_QuantiApp/Features/Tests/RocketDetailTests/RocketDetailCoreTests.swift
@@ -8,7 +8,7 @@ import RocketLaunch
 
 @MainActor
 final class RocketDetailCoreTests: XCTestCase {
-
+  
   func test_data_flow_rocketDetail() async {
     let store = TestStore(
       initialState: RocketDetailCore.State(rocket: .mock),
@@ -17,7 +17,7 @@ final class RocketDetailCoreTests: XCTestCase {
     
     XCTAssertNoDifference(store.state.name, Rocket.mock.name)
     XCTAssertNoDifference(store.state.id, Rocket.mock.id)
-  
+    
     XCTAssertNoDifference(store.state.rocketFirstStage.rocket, store.state.rocket)
     XCTAssertNoDifference(store.state.rocketSecondStage.rocket, store.state.rocket)
     XCTAssertNoDifference(store.state.rocketPhotos.rocket, store.state.rocket)
@@ -25,10 +25,10 @@ final class RocketDetailCoreTests: XCTestCase {
     
     XCTAssertNoDifference(store.state.rocketLaunch, .init())
     
-//    let rocketPhotosAction = RocketPhotosCore.Action.onAppear
-//
-//    store.dependencies.coreMotionClient.xRotationRate = {
-//      try await 3.2
-//    }
+    //    let rocketPhotosAction = RocketPhotosCore.Action.onAppear
+    //
+    //    store.dependencies.coreMotionClient.xRotationRate = {
+    //      try await 3.2
+    //    }
   }
 }
