+++ Ficek_QuantiApp/Ficek_QuantiApp/Features/Tests/RocketListTests/RocketListCoreTests.swift
@@ -11,46 +11,46 @@ import Networking
 @MainActor
 final class RocketListCoreTests: XCTestCase {
   let store = TestStore(
-      initialState: RocketListCore.State(),
-      reducer: RocketListCore()
+    initialState: RocketListCore.State(),
+    reducer: RocketListCore()
   )
   
-    func test_rocket_fetch_successful() async {
-        await store.send(.fetchAsync(.success([Rocket].mock))) {
-          $0.rocketItems = IdentifiedArrayOf(
-                uniqueElements: [Rocket].mock.map { RocketDetailCore.State(rocket: $0) }
-            )
-        }
-        
-      await store.send(.fetchAsync(.failure(NetworkError.timeoutError))) {
-        $0.alert = .errorAlert(error: NetworkError.timeoutError)
-        }
-      
-      await store.send(.alertCancelTapped) {
-        $0.alert = nil
-      }
-      
-      await store.send(.fetchRockets(.failure(.networkError)))
-      
-      await store.send(.fetchRockets(.success([Rocket].mockTest))) {
-        $0.rocketItems = IdentifiedArrayOf(
-          uniqueElements: [Rocket].mockTest.map { RocketDetailCore.State(rocket: $0) }
-        )
-      }
+  func test_rocket_fetch_successful() async {
+    await store.send(.fetchAsync(.success([Rocket].mock))) {
+      $0.rocketItems = IdentifiedArrayOf(
+        uniqueElements: [Rocket].mock.map { RocketDetailCore.State(rocket: $0) }
+      )
     }
+    
+    await store.send(.fetchAsync(.failure(NetworkError.timeoutError))) {
+      $0.alert = .errorAlert(error: NetworkError.timeoutError)
+    }
+    
+    await store.send(.alertCancelTapped) {
+      $0.alert = nil
+    }
+    
+    await store.send(.fetchRockets(.failure(.networkError)))
+    
+    await store.send(.fetchRockets(.success([Rocket].mockTest))) {
+      $0.rocketItems = IdentifiedArrayOf(
+        uniqueElements: [Rocket].mockTest.map { RocketDetailCore.State(rocket: $0) }
+      )
+    }
+  }
   
   func test_data_flow_success() async {
     store.dependencies.rocketClient.fetchAsync = {
       [Rocket].mock
     }
-
-//    store.dependencies.mainQueue = .immediate
+    
+    //    store.dependencies.mainQueue = .immediate
     await store.send(.onAppear)
     
     await store.receive(.fetchAsync(.success([Rocket].mock))) {
       $0.rocketItems = IdentifiedArrayOf(
         uniqueElements: [Rocket].mock.map { RocketDetailCore.State(rocket: $0) }
-    )
+      )
     }
   }
 }
