+++ Ficek_QuantiApp/Ficek_QuantiApp/Infrastructure/Sources/CoreMotionClient/CoreMotionClient.swift
@@ -2,6 +2,6 @@ import Foundation
 import Dependencies
 
 public struct CoreMotionClient {
-   public var xRotationRate: (OperationQueue) async throws -> AsyncThrowingStream<Double, Error>
+  public var xRotationRate: (OperationQueue) async throws -> AsyncThrowingStream<Double, Error>
 }
