+++ Ficek_QuantiApp/Ficek_QuantiApp/Infrastructure/Sources/CoreMotionClient/DependencyValues+coreMotionClient.swift
@@ -2,9 +2,9 @@ import Foundation
 import Dependencies
 
 extension DependencyValues {
-    public var coreMotionClient: CoreMotionClient {
-        get { self[CoreMotionClient.self] }
-        set { self[CoreMotionClient.self] = newValue }
-    }
+  public var coreMotionClient: CoreMotionClient {
+    get { self[CoreMotionClient.self] }
+    set { self[CoreMotionClient.self] = newValue }
+  }
 }
