+++ Ficek_QuantiApp/Ficek_QuantiApp/Infrastructure/Sources/NetworkClientDependency/DependencyValues+networkClient.swift
@@ -3,8 +3,8 @@ import Dependencies
 import Networking
 
 extension DependencyValues {
-    public var networkClient: NetworkClient {
-        get { self[NetworkClient.self] }
-        set { self[NetworkClient.self] = newValue }
-    }
+  public var networkClient: NetworkClient {
+    get { self[NetworkClient.self] }
+    set { self[NetworkClient.self] = newValue }
+  }
 }
