+++ Ficek_QuantiApp/Ficek_QuantiApp/Infrastructure/Sources/NetworkClientDependency/NetworkClient+DependencyKey.swift
@@ -3,16 +3,16 @@ import Networking
 import Dependencies
 
 extension NetworkClient: DependencyKey {
-    public static var liveValue: NetworkClient {
-        return Self(
-            urlRequester: .live(urlSessionConfiguration: .default),
-            networkMonitorClient: .live(onQueue: DispatchQueue.main)
-        )
-    }
-    
-    public static var testValue = NetworkClient(
-        urlRequester: unimplemented("\(Self.self).urlRequester"),
-        networkMonitorClient: unimplemented("\(Self.self).networkMonitorClient")
+  public static var liveValue: NetworkClient {
+    return Self(
+      urlRequester: .live(urlSessionConfiguration: .default),
+      networkMonitorClient: .live(onQueue: DispatchQueue.main)
     )
+  }
+  
+  public static var testValue = NetworkClient(
+    urlRequester: unimplemented("\(Self.self).urlRequester"),
+    networkMonitorClient: unimplemented("\(Self.self).networkMonitorClient")
+  )
 }
