+++ Ficek_QuantiApp/Ficek_QuantiApp/Infrastructure/Sources/UIToolkit/Resources/Image+assets.swift
@@ -7,20 +7,20 @@ import SwiftUI
 // MARK: - Image+
 
 public extension Image {
-    static let burn = Image("Burn")
-    static let engine = Image("Engine")
-    static let fuel = Image("Fuel")
-    static let reusable = Image("Reusable")
-    static let rocketFlying = Image("Rocket Flying")
-    static let rocketIdle = Image("Rocket Idle")
-    static let rocket = Image("Rocket")
-    static let rocketError = Image("rocketError")
+  static let burn = Image("Burn")
+  static let engine = Image("Engine")
+  static let fuel = Image("Fuel")
+  static let reusable = Image("Reusable")
+  static let rocketFlying = Image("Rocket Flying")
+  static let rocketIdle = Image("Rocket Idle")
+  static let rocket = Image("Rocket")
+  static let rocketError = Image("rocketError")
 }
 
 public extension Image {
-    init(_ name: String) {
-        self.init(name, bundle: .module)
-    }
+  init(_ name: String) {
+    self.init(name, bundle: .module)
+  }
 }
 
 public extension Color {
