+++ Ficek_QuantiApp/Ficek_QuantiApp/Infrastructure/Sources/UIToolkit/Resources/LocalizedStringKey+keys.swift
@@ -7,71 +7,71 @@ import SwiftUI
 // MARK: - LocalizedStringKey+
 
 public extension LocalizedStringKey {
-    /// Diameter
-    static let diameter: LocalizedStringKey = "diameter".localizedStringKey
-    /// %lld Engines
-    static func engines(_ p1: Int) -> LocalizedStringKey {
-        "engines %lld".localizedStringKey(with: [p1])
-    }     
-    /// Kann nicht das Data einladen, die Ursache:
-    static let error: LocalizedStringKey = "error".localizedStringKey
-    /// EU metrics
-    static let euMetrics: LocalizedStringKey = "eu_metrics".localizedStringKey
-    /// First flight: %@
-    static func firstFlight(_ p1: String) -> LocalizedStringKey {
-        "first_flight %@".localizedStringKey(with: [p1])
-    }     
-    /// First stage
-    static let firstStage: LocalizedStringKey = "first_stage".localizedStringKey
-    /// Height
-    static let height: LocalizedStringKey = "height".localizedStringKey
-    /// Launch
-    static let launch: LocalizedStringKey = "launch".localizedStringKey
-    /// Launch successful!
-    static let launchSuccessful: LocalizedStringKey = "launch_successful".localizedStringKey
-    /// Mass
-    static let mass: LocalizedStringKey = "mass".localizedStringKey
-    /// Not reusable
-    static let notReusable: LocalizedStringKey = "not_reusable".localizedStringKey
-    /// Overview
-    static let overview: LocalizedStringKey = "overview".localizedStringKey
-    /// Parameters
-    static let parameters: LocalizedStringKey = "parameters".localizedStringKey
-    /// Photos
-    static let photos: LocalizedStringKey = "photos".localizedStringKey
-    /// Reset rocket launch
-    static let resetLaunch: LocalizedStringKey = "reset_launch".localizedStringKey
-    /// Reusable
-    static let reusable: LocalizedStringKey = "reusable".localizedStringKey
-    /// Rockets
-    static let rockets: LocalizedStringKey = "rockets".localizedStringKey
-    /// Second stage
-    static let secondStage: LocalizedStringKey = "second_stage".localizedStringKey
-    /// %lld Seconds burn-time
-    static func secondsBurnTime(_ p1: Int) -> LocalizedStringKey {
-        "seconds_burn_time %lld".localizedStringKey(with: [p1])
-    }     
-    /// Meters remaining to launch %@:
-    static func tiltToLaunch(_ p1: String) -> LocalizedStringKey {
-        "tilt_to_launch %@".localizedStringKey(with: [p1])
-    }     
-    /// %.1lf Tons of fuel
-    static let tonsOfFuelLlf: LocalizedStringKey = "tons_of_fuel %llf".localizedStringKey
-    /// US metrics
-    static let usMetrics: LocalizedStringKey = "us_metrics".localizedStringKey
+  /// Diameter
+  static let diameter: LocalizedStringKey = "diameter".localizedStringKey
+  /// %lld Engines
+  static func engines(_ p1: Int) -> LocalizedStringKey {
+    "engines %lld".localizedStringKey(with: [p1])
+  }
+  /// Kann nicht das Data einladen, die Ursache:
+  static let error: LocalizedStringKey = "error".localizedStringKey
+  /// EU metrics
+  static let euMetrics: LocalizedStringKey = "eu_metrics".localizedStringKey
+  /// First flight: %@
+  static func firstFlight(_ p1: String) -> LocalizedStringKey {
+    "first_flight %@".localizedStringKey(with: [p1])
+  }
+  /// First stage
+  static let firstStage: LocalizedStringKey = "first_stage".localizedStringKey
+  /// Height
+  static let height: LocalizedStringKey = "height".localizedStringKey
+  /// Launch
+  static let launch: LocalizedStringKey = "launch".localizedStringKey
+  /// Launch successful!
+  static let launchSuccessful: LocalizedStringKey = "launch_successful".localizedStringKey
+  /// Mass
+  static let mass: LocalizedStringKey = "mass".localizedStringKey
+  /// Not reusable
+  static let notReusable: LocalizedStringKey = "not_reusable".localizedStringKey
+  /// Overview
+  static let overview: LocalizedStringKey = "overview".localizedStringKey
+  /// Parameters
+  static let parameters: LocalizedStringKey = "parameters".localizedStringKey
+  /// Photos
+  static let photos: LocalizedStringKey = "photos".localizedStringKey
+  /// Reset rocket launch
+  static let resetLaunch: LocalizedStringKey = "reset_launch".localizedStringKey
+  /// Reusable
+  static let reusable: LocalizedStringKey = "reusable".localizedStringKey
+  /// Rockets
+  static let rockets: LocalizedStringKey = "rockets".localizedStringKey
+  /// Second stage
+  static let secondStage: LocalizedStringKey = "second_stage".localizedStringKey
+  /// %lld Seconds burn-time
+  static func secondsBurnTime(_ p1: Int) -> LocalizedStringKey {
+    "seconds_burn_time %lld".localizedStringKey(with: [p1])
+  }
+  /// Meters remaining to launch %@:
+  static func tiltToLaunch(_ p1: String) -> LocalizedStringKey {
+    "tilt_to_launch %@".localizedStringKey(with: [p1])
+  }
+  /// %.1lf Tons of fuel
+  static let tonsOfFuelLlf: LocalizedStringKey = "tons_of_fuel %llf".localizedStringKey
+  /// US metrics
+  static let usMetrics: LocalizedStringKey = "us_metrics".localizedStringKey
 }
 
 public extension String {
-    var nsLocalizedString: String {
-        NSLocalizedString(self, bundle: .module, comment: "")
-    }
-
-    var localizedStringKey: LocalizedStringKey {
-        LocalizedStringKey(nsLocalizedString)
-    }
-
-    func localizedStringKey(with arguments: [CVarArg]) -> LocalizedStringKey {
-        let localizedString = String.localizedStringWithFormat(nsLocalizedString, arguments)
-        return LocalizedStringKey(localizedString)
-    }
+  var nsLocalizedString: String {
+    NSLocalizedString(self, bundle: .module, comment: "")
+  }
+  
+  var localizedStringKey: LocalizedStringKey {
+    LocalizedStringKey(nsLocalizedString)
+  }
+  
+  func localizedStringKey(with arguments: [CVarArg]) -> LocalizedStringKey {
+    let localizedString = String.localizedStringWithFormat(nsLocalizedString, arguments)
+    return LocalizedStringKey(localizedString)
+  }
 }
